+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Implemented enhancements
 - TT-2285 - Add an option to create test users via testapi
 - no ticket - Create user profile on social signup
+- MVP-295 - Expose social profile image
 
 ### Fixed bugs